+++ android/src/main/java/com/hypertrack/sdk/capacitor/HyperTrackCapacitorPlugin.kt
@@ -63,6 +63,11 @@ class HyperTrackCapacitorPlugin : Plugin() {
     }
 
     @PluginMethod
+    fun getWorkerHandle(call: PluginCall) {
+        invokeSdkMethod(SdkMethod.getWorkerHandle, call).toPluginCall(call)
+    }
+
+    @PluginMethod
     fun setIsAvailable(call: PluginCall) {
         invokeSdkMethod(SdkMethod.setIsAvailable, call).toPluginCall(call)
     }
@@ -80,6 +85,11 @@ class HyperTrackCapacitorPlugin : Plugin() {
     @PluginMethod
     fun setName(call: PluginCall) {
         invokeSdkMethod(SdkMethod.setName, call).toPluginCall(call)
+    }
+
+    @PluginMethod
+    fun setWorkerHandle(call: PluginCall) {
+        invokeSdkMethod(SdkMethod.setWorkerHandle, call).toPluginCall(call)
     }
 
     @PluginMethod
@@ -211,6 +221,10 @@ class HyperTrackCapacitorPlugin : Plugin() {
                 HyperTrackSdkWrapper.getName()
             }
 
+            SdkMethod.getWorkerHandle -> {
+                HyperTrackSdkWrapper.getWorkerHandle()
+            }
+
             SdkMethod.locate -> {
                 throw NotImplementedError("Locate is implemented in different way")
             }
@@ -240,6 +254,12 @@ class HyperTrackCapacitorPlugin : Plugin() {
             SdkMethod.setName -> {
                 withArgs<Map<String, Any?>, Unit>(argsJson) { args ->
                     HyperTrackSdkWrapper.setName(args)
+                }
+            }
+
+            SdkMethod.setWorkerHandle -> {
+                withArgs<Map<String, Any?>, Unit>(argsJson) { args ->
+                    HyperTrackSdkWrapper.setWorkerHandle(args)
                 }
             }
         }