+++ ios/Plugin/HyperTrackCapacitorPlugin.m
@@ -12,13 +12,15 @@ CAP_PLUGIN(HyperTrackCapacitorPlugin, "HyperTrackCapacitorPlugin",
            CAP_PLUGIN_METHOD(getLocation, CAPPluginReturnPromise);
            CAP_PLUGIN_METHOD(getMetadata, CAPPluginReturnPromise);
            CAP_PLUGIN_METHOD(getName, CAPPluginReturnPromise);
-           CAP_PLUGIN_METHOD(setIsAvailable, CAPPluginReturnPromise);
-           CAP_PLUGIN_METHOD(setIsTracking, CAPPluginReturnPromise);
-           CAP_PLUGIN_METHOD(setMetadata, CAPPluginReturnPromise);
-           CAP_PLUGIN_METHOD(setName, CAPPluginReturnPromise);
+           CAP_PLUGIN_METHOD(getWorkerHandle, CAPPluginReturnPromise);
            CAP_PLUGIN_METHOD(onSubscribedToErrors, CAPPluginReturnPromise);
            CAP_PLUGIN_METHOD(onSubscribedToIsAvailable, CAPPluginReturnPromise);
            CAP_PLUGIN_METHOD(onSubscribedToIsTracking, CAPPluginReturnPromise);
            CAP_PLUGIN_METHOD(onSubscribedToLocate, CAPPluginReturnPromise);
            CAP_PLUGIN_METHOD(onSubscribedToLocation, CAPPluginReturnPromise);
+           CAP_PLUGIN_METHOD(setIsAvailable, CAPPluginReturnPromise);
+           CAP_PLUGIN_METHOD(setIsTracking, CAPPluginReturnPromise);
+           CAP_PLUGIN_METHOD(setMetadata, CAPPluginReturnPromise);
+           CAP_PLUGIN_METHOD(setName, CAPPluginReturnPromise);
+           CAP_PLUGIN_METHOD(setWorkerHandle, CAPPluginReturnPromise);
 )