+++ src/HyperTrack.ts
@@ -14,6 +14,7 @@ import type { LocationInternal } from './data_types/internal/LocationInternal';
 import type { LocationWithDeviationInternal } from './data_types/internal/LocationWithDeviationInternal';
 import type { Metadata } from './data_types/internal/Metadata';
 import type { OrderHandle } from './data_types/internal/OrderHandle';
+import type { WorkerHandle } from './data_types/internal/WorkerHandle';
 import { registerPlugin } from '@capacitor/core';
 import { Subscription } from './Subscription';
 import { Errors, HyperTrackCapacitorPlugin } from './HyperTrackCapacitorPlugin';
@@ -235,6 +236,17 @@ export default class HyperTrack {
   }
 
   /**
+   * A primary identifier that uniquely identifies the worker outside of HyperTrack.
+   * Example: email, phone number, database id
+   * It is usually obtained and set when the worker logs into the app.
+   * Set it to an empty string "" when the worker logs out of the app to un-bind the device from the worker and
+   * avoid unintentional tracking.
+   */
+  static async getWorkerHandle(): Promise<string> {
+    return hyperTrackPlugin.getWorkerHandle().then((workerHandle: WorkerHandle) => workerHandle.value);
+  }
+
+  /**
    * Requests one-time location update and returns the location once it is available, or error.
    *
    * Only one locate subscription can be active at a time. If you re-subscribe, the old Subscription
@@ -317,6 +329,22 @@ export default class HyperTrack {
       type: 'name',
       value: name,
     } as Name);
+  }
+
+  /**
+   * A primary identifier that uniquely identifies the worker outside of HyperTrack.
+   * Example: email, phone number, database id
+   * It is usually obtained and set when the worker logs into the app.
+   * Set it to an empty string "" when the worker logs out of the app to un-bind the device from the worker and
+   * avoid unintentional tracking.
+   *
+   * @param {string} workerHandle
+   */
+  static setWorkerHandle(workerHandle: string) {
+    hyperTrackPlugin.setWorkerHandle({
+      type: 'workerHandle',
+      value: workerHandle,
+    } as WorkerHandle);
   }
 
   /**