+++ src/HyperTrackCapacitorPlugin.ts
@@ -1,23 +1,22 @@
-import { EVENT_ERRORS, EVENT_IS_AVAILABLE, EVENT_IS_TRACKING, EVENT_LOCATE, EVENT_LOCATION } from "./HyperTrack";
-import { Subscription } from "./Subscription";
-import { Result } from "./data_types/Result";
-import { DeviceId } from "./data_types/internal/DeviceId";
-import { HyperTrackErrorInternal } from "./data_types/internal/HyperTrackErrorInternal";
-import { IsAvailable } from "./data_types/internal/IsAvailable";
-import { IsTracking } from "./data_types/internal/IsTracking";
-import { LocationErrorInternal } from "./data_types/internal/LocationErrorInternal";
-import { LocationInternal } from "./data_types/internal/LocationInternal";
-import { Metadata } from "./data_types/internal/Metadata";
-import { Name } from "./data_types/internal/Name";
+import { EVENT_ERRORS, EVENT_IS_AVAILABLE, EVENT_IS_TRACKING, EVENT_LOCATE, EVENT_LOCATION } from './HyperTrack';
+import { Subscription } from './Subscription';
+import { Result } from './data_types/Result';
+import { DeviceId } from './data_types/internal/DeviceId';
+import { HyperTrackErrorInternal } from './data_types/internal/HyperTrackErrorInternal';
+import { IsAvailable } from './data_types/internal/IsAvailable';
+import { IsTracking } from './data_types/internal/IsTracking';
+import { LocationErrorInternal } from './data_types/internal/LocationErrorInternal';
+import { LocationInternal } from './data_types/internal/LocationInternal';
+import { Metadata } from './data_types/internal/Metadata';
+import { Name } from './data_types/internal/Name';
+import { WorkerHandle } from './data_types/internal/WorkerHandle';
 
 export type Errors = {
-  "errors": HyperTrackErrorInternal[]
-}
+  errors: HyperTrackErrorInternal[];
+};
 
 export interface HyperTrackCapacitorPlugin {
-  addGeotag(
-    ...args: any[]
-  ): Promise<any>;
+  addGeotag(...args: any[]): Promise<any>;
 
   getDeviceId(): Promise<DeviceId>;
 
@@ -33,6 +32,8 @@ export interface HyperTrackCapacitorPlugin {
 
   getName(): Promise<Name>;
 
+  getWorkerHandle(): Promise<WorkerHandle>;
+
   setIsAvailable(isAvailable: IsAvailable): Promise<void>;
 
   setIsTracking(isTracking: IsTracking): Promise<void>;
@@ -41,34 +42,27 @@ export interface HyperTrackCapacitorPlugin {
 
   setName(name: Name): Promise<void>;
 
+  setWorkerHandle(workerHandle: WorkerHandle): Promise<void>;
+
   onSubscribedToErrors(): Promise<void>;
   onSubscribedToIsAvailable(): Promise<void>;
   onSubscribedToIsTracking(): Promise<void>;
   onSubscribedToLocate(): Promise<void>;
   onSubscribedToLocation(): Promise<void>;
 
+  addListener(eventName: typeof EVENT_ERRORS, listenerFunc: (error: any) => void): Subscription;
+
+  addListener(eventName: typeof EVENT_IS_AVAILABLE, listenerFunc: (isAvailable: IsAvailable) => void): Subscription;
+
+  addListener(eventName: typeof EVENT_IS_TRACKING, listenerFunc: (isTracking: IsTracking) => void): Subscription;
+
   addListener(
-    eventName: typeof EVENT_ERRORS, 
-    listenerFunc: (error: any) => void
+    eventName: typeof EVENT_LOCATE,
+    listenerFunc: (locateResult: Result<LocationInternal, HyperTrackErrorInternal[]>) => void,
   ): Subscription;
 
   addListener(
-    eventName: typeof EVENT_IS_AVAILABLE, 
-    listenerFunc: (isAvailable: IsAvailable) => void
-    ): Subscription;
-
-  addListener(
-    eventName: typeof EVENT_IS_TRACKING,
-    listenerFunc: (isTracking: IsTracking) => void
-  ): Subscription;
-
-  addListener(
-    eventName: typeof EVENT_LOCATE, 
-    listenerFunc: (locateResult: Result<LocationInternal, HyperTrackErrorInternal[]>) => void
-  ): Subscription;
-
-  addListener(
-    eventName: typeof EVENT_LOCATION, 
-    listenerFunc: (location: Result<LocationInternal, LocationErrorInternal>) => void
+    eventName: typeof EVENT_LOCATION,
+    listenerFunc: (location: Result<LocationInternal, LocationErrorInternal>) => void,
   ): Subscription;
 }