+++ CHANGELOG.md
@@ -5,6 +5,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.4.0] - 2024-06-13
+
+### Added
+
+- New `setWorkerHandle` and `getWorkerHandle` can be used to identify workers
+  - We observed our customers identify worker devices via `HyperTrack.metadata`, so we decided to make it a first class citizen in our API.
+  - If you previously used `metadata` to identify workers, we suggest using `workerHandle` for this purpose instead.
+
+### Changed
+
+- Updated HyperTrack SDK iOS to [5.6.0](https://github.com/hypertrack/sdk-ios/releases/tag/5.6.0)
+- Updated HyperTrack SDK Android to [7.6.0](https://github.com/hypertrack/sdk-android/releases/tag/7.6.0)
+
 ## [3.3.1] - 2024-05-24
 
 ### Changed
@@ -215,3 +228,4 @@ We are excited to announce the release of HyperTrack Ionic Capacitor SDK 2.0.0, 
 [3.2.2]: https://github.com/hypertrack/sdk-ionic-capacitor/releases/tag/3.2.2
 [3.3.0]: https://github.com/hypertrack/sdk-ionic-capacitor/releases/tag/3.3.0
 [3.3.1]: https://github.com/hypertrack/sdk-ionic-capacitor/releases/tag/3.3.1
+[3.4.0]: https://github.com/hypertrack/sdk-ionic-capacitor/releases/tag/3.4.0